requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
cssselect==1.2.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
magic-filter==1.0.12
multidict==6.0.4
packaging==23.2
parsel==1.8.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.7
w3lib==2.1.2
yarl==1.9.2
